+++ libs/core/serialization/include/hpx/serialization/input_archive.hpp
@@ -74,7 +74,7 @@ namespace hpx::serialization {
             flags_ = flags;
 
             // load the zero-copy limit used by the other end
-            std::size_t zero_copy_serialization_threshold;
+            std::uint64_t zero_copy_serialization_threshold;
             load(zero_copy_serialization_threshold);
             buffer_->set_zero_copy_serialization_threshold(
                 zero_copy_serialization_threshold);
